% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pbounds.R
\name{Pbounds}
\alias{Pbounds}
\title{Calculation of lower and upper probability bounds}
\usage{
Pbounds(target, ...)
}
\arguments{
\item{target}{The desired probability expression.}

\item{...}{Probability constraints (see Details).}
}
\value{
A vector of \code{min} and \code{max} values of the desired probability.
}
\description{
Compute the minimum and maximum value that the probability of a logical formula conditional on another one, given numerical or equality constraints for the conditional probabilities of other logical formulae.
}
\details{
This function takes as first argument the desired probability expression, and as subsequent arguments the known probability constraints.

The constraints can be of the form \verb{P(...|...) == [number]} or \code{P(...|...) == P(...|...)}, even with inequalities instead of equalities.

All probability expressions must start with \code{P} or \code{p} or \code{Pr} followed by parentheses. The symbol for the conditional bar and the logical connectives are as follows:
\itemize{
\item Conditional bar: \code{~}
\item Not: \code{!}
\item And: \code{&&} or \code{&}
\item Or: \code{||} or \code{|}
\item If-then: \verb{\%>}
}
}
\examples{
Pbounds(
  target = P(a & b ~ h), # P(a ∧ b | h) in standard notation
  P(a ~ h) == 0.3,
  P(b ~ h) == 0.6
)
## min max
## 0.0 0.3

Pbounds(
    target = P(a & b ~ h), # P(a ∧ b | h) in standard notation
    P(a ~ h) == 0.3,
    P(b ~ a & h) == 0.2
)
##  min  max
## 0.06 0.06

## Solution to the Monty Hall problem:
Pbounds(
  target = P(car1 ~ you1 & host2 & I),
  P(car1 | car2 | car3 ~ I) == 1,
  P(car1 & car2 ~ I) == 0,
  P(car1 & car3 ~ I) == 0,
  P(car2 & car3 ~ I) == 0,
  P(host1 & you1 ~ I) == 0,
  P(host2 & you2 ~ I) == 0,
  P(host3 & you3 ~ I) == 0,
  P(host1 & car1 ~ I) == 0,
  P(host2 & car2 ~ I) == 0,
  P(host3 & car3 ~ I) == 0,
  P(host1 & host2 ~ I) == 0,
  P(host1 & host3 ~ I) == 0,
  P(host2 & host3 ~ I) == 0,
  P(host1 | host2 | host3 ~ I) == 1,
  P(host2 ~ car1 & you1 & I) == P(host3 ~ car1 & you1 & I),
  P(car1 ~ I) == 1/3,
  P(car2 ~ I) == 1/3,
  P(car3 ~ I) == 1/3,
  P(car1 ~ you1 & I) == 1/3,
  P(car2 ~ you1 & I) == 1/3,
  P(car3 ~ you1 & I) == 1/3
)
##      min      max
## 0.333333 0.333333

}
